Optimize listJar() method
listJar() is a method that returns a list of files
contained within a Jar file according to a glob pattern.

It used to rely on a ZipInputStream that needs to read
the entire file to discover its content.

Now switched to a ZipFileSystem that is able to use the
Central Directory at the end of the file directly to list
contained files, therefore it is by order of magnitude
faster.

Change-Id: Idecd9c512fc878d13e04f60d9500173b26220db0